requirements.txt
astroid==3.2.4
beautifulsoup4==4.12.3
bs4==0.0.2
coverage==7.6.1
dill==0.3.8
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pylint==3.2.6
pytest==8.3.2
PyYAML==6.0.2
soupsieve==2.6
tomlkit==0.13.2
